RFC: constify Cython output all over the place  (newbie approach)
You know, when developing code, it is very tedious to look for meaningful
errors and warnings in presence of tons of noise like

    warning: deprecated conversion from string constant to ‘char*’

And you know what? It seems in the next version of gcc, this deprecation
warnings will be turned into errors.

----

Python sources already use 'const' keyword freely, so I think it's time to add
constify bits all over the place.