Emit code to guard declarations for this module's public functions.
It appears that the machinery is all there to protect the declaration
of inconsistent linkage types (i.e. dllimport vs. dllexport) by using
a guard unique to the pyx file.  This patch simply emits that guard
for the current module to prevent the error "inconsistent linkage types"
on a Windows-based system.