Stopgap support for numpy 1.7
NumPy is starting to seriously deprecating access to the member fields
in an ndarray (it was always frowned upon, but now it is starting to
become enforced). To support the large body of Cython code out there
accessing these fields (arr.shape[0] and so on), we special-case
PyArrayObject in Cython, with special knowledge of the NumPy API.

Ideally, we may introduce features in Cython in the future that allows
specifying this kind of magic with syntax in pxd files, and then we can
move away from special-casing NumPy.