Fix computation of destructor name with namespaces
In the computation of the destructor name,
currently the class name is computed by taking
everything at the right of the last ::
to remove the namespaces.
This is incorrect with template arguments containing :: ,
e.g. if the class is
Namespace1::Namespace2::myclass < A::B::C, D< F::G> >
Correction is to take first the part before the <.