Change __Pyx_TraceDeclarations to be C89 compliant
I added a line note elsewhere about compiling a memoryview-heavy file with profiling enabled and getting an error from MSVC about how C89 doesn't let you declare more variables after an if statement.

I had to change $frame_code_cname to a non-static local as a static local does not allow initializing with a non-constant. I can't tell why it needed to be static, as the assignment runs during every function call.

Just having the if statement in a `Declarations` macro kind of makes it a misnomer. Also, that if statement was fairly iffy in and of itself - why would you need to check if `codeobj` is NULL if you're just going to assign NULL anyway if it is?