Change temp type used for buffer indexing to ssize_t
Changes the type of the index temps in IndexNode.buffer_lookup_code.

This matches the type of MemoryView.shape and ndarray.shape, and fixes warnings (and possible bugs) related to truncating the index value with loss of data when the type of the index expression is smaller than a ssize_t.