setup.py shouldn't write arbitrarily to stdout (especially when passed display options)
In general it's best for actions performed in `setup.py` to be deferred until the command to which they are relevant is actually run.  For example most of what `compile_cython_modules` does should be deferred until finalizing the `build_ext` command.  But as a simpler workaround, just ensure that anything it prints is to stderr.

Otherwise this breaks certain assumptions, such as that `./setup.py --name` will print (on stdout) the distribution name (and nothing else), and likewise for `./setup.py --version`.