add ext_opts parameter to cythonize (fixes #1480)
Trying to set default `Extension()` options by passing them to `cythonize()`
is a common mistake, being made even in the Cython docs (see #1480). This patch
adds the capability to append to the generated Extension options by passing a
dictionary as parameter `ext_opts` to `cythonize()`, and updates the documtentation
accordingly.