Suppress unused function warnings for Clang.
Cython uses many static inline utility functions.  Most compilers
don't emit warnings for such functions, but Clang does if they're
not defined in header files.  As such functions are often defined
as a group (e.g. utility code for a specific type) it would be
cumbersome to thin them out to those that happen to actually get
used.  Instead, simply suppress this warning for this compiler
for all inline functions.