Add zip_safe=False to userguide setup.py
This adds notes in the documentation to include zip_safe=False when
using setuptools instead of distutils for setup.py.

Because cimport will not load pxd definitions from a zipped egg file and
the default format when installing from a source package using
setuptools and "python setup.py install" is to create a zipped egg,
including this option will guard against this error when people install
this way.