Call out the different name used for the wrapper
Explicitly call out the different name used for the wrapper `.pyx` file vs. the other C++ sources

As a new user, I was following along and just called the `.pyx` file `Rectangle.pyx` like everything else which was kinda confusing then when in the `wrapping_CPlusPlus/setup.py` there's a `setup(ext_modules=cythonize("rect.pyx"))` and I was thinking that maybe that file was magically created somehow and was miffed.

Hoisted the blurb on why the file name needs to be different without some additional work and worded it in there.