Search first for adjacent pxd in compile
In the function DependencyTree.cimported_files we will find a .pxd file
in the same directory as our .pyx file regardless of the assigned
package name; but when we go on to compile same file this file is not
found; we only look for file matching the assigned package name.

In this change I first find src/bar.pxd matching src/bar.pyx even when
the resulting package would be foo.bar.