include: posix.mman += mlock2, MLOCK_ONFAULT, MCL_ONFAULT
mlock2 is Linux-specific version of mlock that accepts flags. The only
flag available so far is MLOCK_ONFAULT which asks to lock pages only at
the time when they are faulted in, not in the beginning. MCL_ONFAULT is
Linux-specific flag to mlockall which requests similar behaviour.

http://man7.org/linux/man-pages/man2/mlock2.2.html

(cherry picked from commit 83a5700b30cdcb538ec8c23b4fcff1db54449abc)