3057-include-once: protect cached_function and cached_method against recursion.
A recursion call in a cached_function or a cached_function is not
expected to happen.
In that case, it is best to raise an exception so the caller can act in
consequence.
For example, when parsing dependencies, we can have a recursive
dependency. In that case, it is best to ignore it.