3057-include-once: implement 'include_once' statement.
The 'include_once' statement behave like the 'include' statement, but
the included file is guarantee to be included only once per module.

To ensure this, the parser is tracking every included file paths and
adds them in new set variable `included_file_paths`.
We don't use the existing `included_files` because it stores the names
of the included files and it is not backward compatible to change the
types and the content of this variable.