Accept unicode in identifiers (step 1)
This only accepts it from a code matching point of view - it
tends to break if you actually use it in Cython. However existing
(ascii only) is still handled OK by this.

It's possible this should turn itself on and off appropriately
depending on version and encoding however there isn't anywhere
obvious to do it