Allow unicode named structs/other C items
When they are to be used internally the name is mangled with
punycode as normal. When they are to be used externally
(e.g. "cdef public" or "cdef from extern") the name is taken
exactly as-is and simply slash-escaped ("\uNNNN"). The
vast majority of C compilers are capable to dealing with
\uNNNN characters in literal names.