Fixed generator expression scope (hopefully)
The basic outline of moving the variables to a temp remains as
before. Now after the loop optimizatation step any temp variables
that appear both inside and outside the generator expression
are converted to arguments to the genexpr function. They can
then be captured as closure variables by the normal rules
rather than creating a special mechanism