Added fastcall_args directive
This turns *args and/or **kwds into a special fastcalltuple
or fastcalldict type, which may be possible to use more
efficiently (but are more limited in their usage).

They're enabled by an `@cython.fastcall_args("*"/"**"/"both")
decorator on the relevant function.

Currently `*a` is generally a performance win. `**` is
less convincing, but often better. Iterating over `kwds.items()`
is a notable regression.

The main missing optimization at the moment would be being able
to to use these types to call into other functions efficiently.