Consider builtin types as opaque when importing in limited API
cython supports a construct like this to declare members of object types:
```
ctypedef class __builtin__.py_long [object PyLongObject]:
    cdef digit* ob_digit
```

The generated code checks that `tp_basicsize` matches the size of
corresponding type struct. This does not work in the limited API in
which builtin type structs are opaque; this disables the test in that
case.

This helps code that contains such a `ctypedef` but does not otherwise
use the type.

Based on Facebook D20344150