Dropped unused code-paths associated with if cython.compiled early
This allows things to work like:
    # in pxd file
    from libc.math cimport sin

    # in py file
    if not cython.compiled:
        from math import sin  # previously failed with cython compile
           # error because it was assigning to a cdef name

This seems worthwhile because it makes it easier to generate code
that re-assigns cdef names and so works in both modes.

For the sake of simplicitity it only handles very simple cases
and doesn't attempt a full exercise in constant folding at this
stage (hence `if not cython.compiled and True` isn't handled).
This seems like a sensible compromise.