FIX: Ignore unused ctuples (GH-3543)
generate_type_definitions maintained a list of ctuples for which code
was already generated, but the logic did not match the logic used in
generate_type_header_code, which does not generate code if the entry is
marked as unused.

If the first ctuple of a certain type was marked unused, but later
instances were not, this could result in no code being generated for
the ctuple at all. Resolve by verifying whether the ctuple entry is used
before adding it to the list.