Increment result of module init function
Summary: Fix problem where the code generated for the limited API would not increment the module before returning it from the PyInit function (and decref the temporary that held the result of `PyModule_Create`). Fix this by keeping the module in the temporary throughout the function and using it for the `return` as well.

Based on Facebook D21894032