Refactor string constant handling
Instead of generating dozens of individual variables for string
constants, create an enum with indexes and an array.

This has the benefit that we no longer need to put a variable reference
into each string table entry. This in turn allows us to use the same
code regardless of `CYTHON_USE_MODULE_STATE`.