Made C++ temps a special type to avoid default constructor
This uses placement new to avoid requiring a default constructor
(at the cost of using a little extra space for an "assigned" bool)
The pattern is pretty similar to std::optional. This gives a little
extra flexibility in terms of the C++ classes that Cython can handle.

(built on top of https://github.com/cython/cython/pull/3792
for simplicity, but needn't require that - it only requires a change
to some #if lines)

Longer-term it might be possible to support something similar for
regular cdef variables, but it'd probably need some extra syntax
since I doubt we'd want to change the default behaviour.