Fix bug in cpp iteration over rvalue-dependant attribute
The implementation of the Python-style `for ... in` loop over an
iterable cpp sequence (using `begin()` and `end()`) needs to determine
whether the sequence should be stored in a temporary variable or
if it can be iterated over directly.

The previous code assumed that attributes are always safe to be
iterated directly, but attributes can depend on an rvalue, as in:

    for i in returns_object_with_vector_attribute().vector:
        ...

In such a case we need to store the `in` operand in a temporary value
on which `begin()` and `end()` can be safely called.

It might make sense to move the test into tests/run/cpp_iterators