Fix cpp iteration over simple expression depending on temporaries
The implementation of the Python-style `for ... in` loop over an
iterable cpp sequence (using begin() and end()) needs to determine
whether the sequence should be stored in a temporary variable or
if it can be iterated over directly.

A previous fix considers that the sequence can be accessed directly
if `sequence.is_simple()` evaluates to True.

However it is at least theoretically be possible for an expression
to both be simple and to depend on temporary subexpressions. In that
case, because the CppIteratorNode is always a temporary, the compiler
will by default eagerly free the temporary subexpressions for reuse
after the result of the sequence has been evaluated the first time.
This will result in subsequent calls to `result()` returning `None`
even though the required variables are still available and in scope
in the generated C++ code.

This commit postpones the release of temporary subexpressions until
the CppIteratorNode is itself disposed-of when the sequence is
accessed directly.