Add code to warn user about failed cdef class lookups
One of the most commonly seen user errors is to try to access
cdef methods/attributes of a cdef class when Cython doesn't
know the type.

This attempts to add warnings to alert the user so that they
can hopefully correct it themselves. It obviously isn't
(and can't be) comprehensive but does catch the obvious cases.

It's possible that it'll be a bit noisy/over-sensitive.
Hopefully the test-suite should give an indication of how
many false positives it generates.