Implement PEP560 inheritance (__mro_entries__)
Fixes https://github.com/cython/cython/issues/3537

Both the C code and the tests are largely copied from CPython
so hopefully should be pretty reliable.

This doesn't backport perfectly so I've hard to disable a few
tests of Python 2 (and may have to on earlier Python 3 versions too).
I think that's OK.

I haven't applied this at all to cdef classes. Maybe it could be
applied to the second+subsequent bases but I don't think it's
needed for the initial implementation.