Implement PEP-560 inheritance (__mro_entries__) (GH-4005)
Fixes https://github.com/cython/cython/issues/3537

Both the C code and the tests are largely copied from CPython.

Note that this currently only applies to Python classes, not to cdef classes. Maybe it could be applied to the second+subsequent bases but I don't think it's needed for the initial implementation.