Fix windows linkage error for unicode modules by providing a dummy function (GH-4125)
A bug in distutils means that it tries to export a slightly incorrectly named PyInitU function. The fix in https://bugs.python.org/issue39432 wasn't quite right (`"_name".encode("punycode") != "name".encode("punycode")`) and thus the "unicode_imports" test is failing on Windows 3.8+ when distutils explicitly tries to export a non-existing symbol.

We now create a dummy function with that alternative name so that the export doesn't fail.