Use std::move in c++ during yield context switch.
When compiling pyarrow with cython 3.0 we get an error about the copy
constructor of an object has been deleted on the generated context switch code.

This PR fixes it.

I'd like to add a unit test, but would need some input on which folder
or file under tests is a good place for it.