Allow releasing gil in nogil functions
Fixes #4137

Also adds a check whether we have the GIL before doing so. This
is important because Py_UNBLOCK_THREADS is documented as unsave
if we don't hold the GIL. This bit of it is possibly a candidate to be
backported (since the pattern was allowed in Cython 0.29 but
the behaviour looks like it could be quite buggy).