Changes needed to begin testing GraalPython
Very little works at present, there looks to be huge memory-leaks,
and the test-suite runs very slowly. It definitely isn't worth adding
anything on Github Actions at this stage.

It's just designed to get over the initial hurdle so that it
can be tested and improvements can be made if anyone is
inlined.