Help type checkers find Shadow typings
Before this commit there was certain issues where static type checkers
had issues finding the typings in Shadow.pyi and it was also not always
included in installations.

This commit additionally poses as a necessary renovation of the typings
themselves by adding some missing members and switching to newer synta
because .pyi files are not constrained to a specific Python version.