Check for dodgy multiple inheritance of cdef class
Closes https://github.com/cython/cython/issues/4350 (by raising
a runtime error on the cases that will fail).

Some notes:

* It only works on Python 3.6+. I think that's OK for something
that's designed to help users not to shoot themselves in the
foot rather than a core feature. Supporting earlier versions
would probably be more intrusive.

* The test case is in "errors" rather than "run" even though its
checking for runtime errors. I decided that was the most
suitable place, but maybe I'm wrong...

* It's interesting that Python is more lenient on multiple
inheritence of cdef classes that Cython would be if "C" were
defined as cdef classes

* Some of these errors could be detected at compile-time in
principle too. This seems like a lot of effort given that we'll
need the runtime check anyway. Thus I haven't done it.