Fix fused cpdef default arguments
A couple of things were going wrong:
* they're creating CloneNodes (but not requiring the contents
of the clone of the clone node to be temp)
* Assignment from a clone node generates cleanup code (which
is against the general rules of a clone node), and also loses
a reference via giveref
* cpdef functions cause a small memory leak (#4412) by
assigning to their constants twice. This is unfortunately
difficult to test for. With this patch we no longer leak, but
still duplicate a little bit of work.