Try to fix exception when debugging a module with linetrace enabled
With linetrace enabled trying to debug with VS Code (Which uses the
pydevd debugger internally), causes a "TypeError: 'NoneType' object is
not callable" due to mishandling of the f_trace member of frame objects

f_trace is an internal member used by trace_trampoline to support a
Python trace function that returns a different trace function than
itself, basically the thread local trace function. Cython is setting it
to None, but it should (Surprisingly and unlike other PyObject*) be
`NULL` or a callable `PyObject` but not `None. Actually, Cython should
just avoid touching the it at all, as it is only used by
trace_trampoline internally for it's own operation and isn't supposed to
be touched by code externally calling the c_tracefunc (Which again is a
Python implementation detail, FUN!).

Now f_trace might have be set to None for some other purpose, but it
shouldn't really be touched by Cython.

Fixes #4443