Only set HAVE_VECTORCALL flag when using vectorcall (GH-4453)
Before this the _Py_TPFLAGS_HAVE_VECTORCALL was set all the time
when using the limited api, but a CyFunction never sets up a
vectorcall function. This caused basically all function calls to
crash when binding=True with the limited API.

Also removes the premature refnanny usage when setting up the module.