Disable Cython reimplementation of trashcan
If I understand correctly, the Cython-specific reimplementation
of trashcan isn't needed of Python 3.8+ (because the patch it
was based on was merged into Python). Therefore on these versions
it can be disabled and the Python macros used instead.

This is specifically needed for Python 3.11 since it removes
hides functions that are used in the Cython reimplementation.
However, I've gone back further.

(Admittedly I don't understand the exact special-case this was for
too well, but I'm hoping this is tested...)